requirements.txt
Babel==2.7.0
beautifulsoup4==4.8.1
boto==2.49.0
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-jsonfield==1.3.1
django-markdown==0.8.4
django-mdeditor==0.1.16
django-phonenumber-field==3.0.1
django-picklefield==2.0
django-sendgrid==1.0.1
django-sendgrid-v5==0.8.0
django-sendsms==0.4
django-ses==0.8.12
django-sms-gateway==1.2.1
django-smtp-ssl==1.0
django-twilio==0.11.0
django-user-messages==0.6.2
django-uuidfield-2==0.6.6
django-widget-tweaks==1.4.5
future==0.18.1
gunicorn==19.9.0
idna==2.8
Markdown==3.1.1
martor==1.4.4
ngrok==0.0.1
phonenumbers==8.10.21
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
PySocks==1.7.1
python-decouple==3.1
python-http-client==3.2.1
pytz==2019.1
requests==2.22.0
sendgrid==6.1.0
sendgrid-django==4.2.0
six==1.12.0
soupsieve==1.9.4
South==1.0.2
sqlparse==0.3.0
twilio==6.7.0
urllib3==1.25.6
whitenoise==4.1.4
zerosms==0.0.1
